requirements.txt
beautifulsoup4==4.12.2
billiard==4.1.0
boto3==1.26.119
botocore==1.29.165
bs4==0.0.1
celery==5.3.1
certifi==2023.7.22
crochet==2.0.0
Flask==2.2.2
Flask-Cors==3.0.10
google-ads==22.0.0
google-api-core==2.10.1
google-api-python-client==2.36.0
google-auth==2.22.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
gunicorn==21.2.0
lxml==4.9.2
mysql-connector-python==8.0.30
mysqlclient==2.2.0
numpy==1.24.4
oauthlib==3.2.2
pandas==1.3.5
PyMySQL==1.0.3
python-dotenv==1.0.0
requests==2.31.0
Scrapy==2.5.1
SQLAlchemy==1.4.43
urllib3==1.26.16
